Created class for ConsoleOutput and moved printTable Method to new class
