+++ src/test/java/hwr/oop/budgetbook/ReadWriteCsvTest.java
@@ -9,9 +9,10 @@ public class ReadWriteCsvTest {
     void readCsvFile_TryIfATestFileIsReadAndDisplayed() {
         String path = "C:\\Users\\Malte\\IdeaProjects\\hwr-oop-project-bank\\src\\test\\java\\hwr\\oop\\budgetBook\\test.csv";
         ReadWriteCsv test = new ReadWriteCsv();
+        ConsoleOutput output = new ConsoleOutput();
         try {
             String[][] data = test.readCsvFile(path);
-            test.printTable(data);
+            output.printTable(data);
         } catch (Exception e) {
             e.printStackTrace();
             System.out.println("Error");
@@ -19,13 +20,15 @@ public class ReadWriteCsvTest {
     }
 
     @Test
-    void writeCsvFile_WriteACsvFileWithATestDataset(){
-        String[] dataset = new String[] {"doe78","1234","John","Doe"};
+    void writeCsvFile_WriteACsvFileWithATestDataset() {
+        String[] dataset = new String[]{"doe78", "1234", "John", "Doe"};
         String path = "C:\\Users\\Malte\\IdeaProjects\\hwr-oop-project-bank\\src\\test\\java\\hwr\\oop\\budgetBook\\test1.csv";
 
         ReadWriteCsv test = new ReadWriteCsv();
+        ConsoleOutput output = new ConsoleOutput();
+
         try {
-            test.writeDataset(dataset,path);
+            test.writeDataset(dataset, path);
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -35,7 +38,7 @@ public class ReadWriteCsvTest {
         } catch (Exception e) {
             e.printStackTrace();
         }
-        test.printTable(data);
+        output.printTable(data);
 
     }
